[move] quest>component 패키지 이동
